게시물 글쓰기 의사코드 추가

+++ community-server/src/service/board.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { domains } from '../../../domains';
@@ -9,17 +9,33 @@ import { BoardSearchDto } from '../dto/boardSearchDto';
 export class BoardService {
   constructor(
     @InjectRepository(domains.Board)
-    private readonly boardRepository: Repository<domains.Board>,
+    private readonly boardRepository: Repository<domains.Board>, //private readonly userRepository: Repository<domains.User>, //private readonly doctorRepository: Repository<domains.Doctor>,
   ) {}
 
-  createBoard(boardDto: BoardDto) {
+  createBoard(boardDto: BoardDto, user) {
+    if (!user) {
+      throw new HttpException(
+        '로그인 상태가 아닙니다.',
+        HttpStatus.UNAUTHORIZED,
+      );
+    }
+    boardDto.userId = user.id;
     return this.boardRepository.save(this.boardRepository.create(boardDto));
   }
 
-  updateBoard(id: number, boardDto: BoardDto) {
+  updateBoard(id: number, boardDto: BoardDto, user) {
+    if (!user) {
+      throw new HttpException(
+        '로그인 상태가 아닙니다.',
+        HttpStatus.UNAUTHORIZED,
+      );
+    }
+
+    if (id !== user.id) {
+      throw new HttpException('글쓴이가 아닙니다.', HttpStatus.UNAUTHORIZED);
+    }
     const board = this.findOne(id);
-    board.userId = boardDto.userId;
-    board.userName = boardDto.userName;
+    // board.userid 가 로그인한 user.id랑 맞는지 체크한다.
     board.title = boardDto.title;
     board.content = board.content;
     return this.boardRepository.update(id, board);
@@ -54,7 +70,7 @@ export class BoardService {
       board = this.boardRepository
         .createQueryBuilder('board')
         .leftJoinAndSelect('board.user', 'user')
-        .select(['board', 'user.username']);
+        .select(['board', 'user.name']);
 
       board =
         boardSearchDto.title !== 'undefined'
@@ -63,9 +79,9 @@ export class BoardService {
             })
           : board;
       board =
-        boardSearchDto.username !== 'undefined'
-          ? board.where('user.username like :username', {
-              username: '%' + boardSearchDto.username + '%',
+        boardSearchDto.name !== 'undefined'
+          ? board.where('user.name like :name', {
+              name: '%' + boardSearchDto.name + '%',
             })
           : board;
 
@@ -75,15 +91,23 @@ export class BoardService {
     } catch (error) {
       return null;
     }
-
-    //return this.boardRepository.find();
   }
 
   findListByCategory(category: string) {
     return this.findByCategory(category);
   }
 
-  deleteBoard(id: number) {
+  deleteBoard(id: number, user) {
+    if (!user) {
+      throw new HttpException(
+        '로그인 상태가 아닙니다.',
+        HttpStatus.UNAUTHORIZED,
+      );
+    }
+
+    if (id !== user.id) {
+      throw new HttpException('글쓴이가 아닙니다.', HttpStatus.UNAUTHORIZED);
+    }
     return this.boardRepository.delete(id);
   }
 }
